BAH-4858 | Refactor. groupObservationsByEncounter to observationService
