+++ packages/bahmni-services/src/observationService/__mocks__/observationMocks.ts
@@ -197,3 +197,115 @@ export const mockFormsEncounter = {
     },
   ],
 };
+
+export const mockObservationWithFormPath: Observation = {
+  resourceType: 'Observation',
+  id: 'obs-1',
+  status: 'final',
+  code: {
+    coding: [
+      {
+        code: '5087AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+        display: 'Pulse',
+      },
+    ],
+    text: 'Pulse',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  encounter: {
+    reference: 'Encounter/encounter-1',
+  },
+  effectiveDateTime: '2026-07-20T09:59:41+00:00',
+  valueQuantity: {
+    value: 76,
+    unit: 'beats/min',
+  },
+};
+
+export const mockObservationsForVitals: Observation[] = [
+  mockObservationWithFormPath,
+  {
+    ...mockObservationWithFormPath,
+    id: 'obs-vitals-2',
+    code: {
+      coding: [
+        {
+          code: '5090AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+          display: 'Height (cm)',
+        },
+      ],
+      text: 'Height (cm)',
+    },
+    valueQuantity: {
+      value: 12.0,
+      unit: 'cm',
+    },
+  },
+];
+
+export const mockEncounterWithProvider: Encounter = {
+  resourceType: 'Encounter',
+  id: 'encounter-1',
+  status: 'unknown',
+  class: {
+    system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode', // NOSONAR
+    code: 'AMB',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  participant: [
+    {
+      individual: {
+        reference: 'Practitioner/60b31d2a-1d0c-11f1-b099-5a3ed7acdb7e',
+        type: 'Practitioner',
+        display: 'Super Man',
+      },
+    },
+  ],
+  period: {
+    start: '2026-07-20T09:59:41+00:00',
+  },
+};
+
+export const mockEncounterWithoutProvider: Encounter = {
+  resourceType: 'Encounter',
+  id: 'encounter-2',
+  status: 'unknown',
+  class: {
+    system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode', // NOSONAR
+    code: 'AMB',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  participant: [],
+  period: {
+    start: '2026-07-20T10:00:00+00:00',
+  },
+};
+
+export const mockEncounterWithoutPeriodStart: Encounter = {
+  resourceType: 'Encounter',
+  id: 'encounter-3',
+  status: 'unknown',
+  class: {
+    system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode', // NOSONAR
+    code: 'AMB',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  participant: [
+    {
+      individual: {
+        reference: 'Practitioner/another-practitioner',
+        type: 'Practitioner',
+        display: 'Dr. Smith',
+      },
+    },
+  ],
+  period: {},
+};
