+++ packages/bahmni-services/src/observationService/index.ts
@@ -3,4 +3,6 @@ export {
   getPatientObservationsWithEncounterBundle,
   getPatientObservations,
   getObservationsBundleByEncounterUuid,
+  groupObservationsByEncounter,
 } from './observationService';
+export type { EncounterGroup } from './observationService';
